MAINT: pin versions of all build dependencies
The argument for pinning build dependencies in this repository is:

The point is to be able to build wheels in a controlled way that we
can be sure works.

Testing whether there are incompatibilities in build dependencies (e.g.
whether latest Cython version works) on the other hand should be done
in the main CI repository.