Set LDFLAGS for stripping at link time
Numpy distutils appears to only look at the LDFLAGS environment variable
when determining how to link fortran.

Note that distutils and numpy.distutils interpret the LDFLAGS variable
in slightly different ways, with the latter requiring that -shared is
included there. To work around this, we concatenate the Python and numpy
defaults.